feat(v3.0-1c/1d/1e): finish plugin runtime + fix client-bundle blocker
Completes Phase 1 of the v3.0 plugin runtime (#55). Builds on the WIP
commit with the fixes surfaced by the bug hunt.

## BLOCKER fix: plugin nav vanishes after client hydration

Bug hunt found: `src/plugins/hello/index.ts` was imported only via
runtime.ts → hooks.server.ts (server-only chain). Sidebar.svelte runs
in the browser bundle, imports its own instance of sidebar-nav.ts,
which had zero plugin registrations. SSR HTML briefly showed the
plugin entry; hydration wiped it.

Fix: new $lib/plugins/registrations.ts imports every enabled plugin
for side effects. sidebar-nav.ts imports it at bottom (client bundle
picks it up); runtime.ts also imports it (server bundle). Both bundles
now run the same module-load registrations.

## Client-bundle server-import blocker

registerWebhookEvent lived in $lib/server/content/types.ts. Plugins
importing it dragged server-only code into the client bundle, which
SvelteKit rejects.

Fix: extracted KnownWebhookEvent, WebhookEvent, registerWebhookEvent,
listKnownWebhookEvents, WEBHOOK_EVENTS into new
$lib/plugins/webhook-events.ts (pure ES module, no server deps).
types.ts re-exports them (values + types) so existing imports keep
working.

## Other fixes from the hunt

- runtime.ts initPromise: try/finally guarantees `initialized = true`
  even on unexpected outer throws (was: could pin future requests to
  a rejected promise)
- hello/index.ts: registrations moved from onInit() → module load
  (runs before first request, ensuring first-render correctness)
- svelte.config.js: added $plugins alias for src/plugins
- hello/+page.server.ts: replaced fragile "../../../../plugins/..."
  relative import with $plugins/hello/schema
- hooks.server.ts: pluginInitHook comment corrected — most plugins
  don't need onInit; sidebar/webhook registration is at module load

## Docs shipped

- docs/plugin-authoring.md: contract, file layout, module-load
  registration convention, migration + route + audit + webhook
  patterns, uninstall/data-retention policy, roadmap for v3.5 CLI
- docs/PLUGINS.md: curated list — @khaopad/plugin-hello (in-tree v3.0),
  @khaopad/plugin-shop planned v3.1, @khaopad/plugin-reviews v3.4

## Deliberately deferred (v3.5)

- pnpm khaopad plugin list|install|remove CLI
- khaopadCompat peer-dep check at install time
- npm package discovery via package.json's khaopad.plugins field
- Plugin settings cards in /admin/settings
- Automated acceptance test suite (blocks on #63 vitest baseline)

## Verify

- pnpm run check: 0 new errors (3 pre-existing paraglide errors from #62)
- pnpm run build: passes in 8s (Cloudflare adapter output emitted)

Closes: v3.0 plugin runtime (#55)
