+++ src/lib/plugins/runtime.ts
@@ -10,13 +10,20 @@
  * `khaopad.plugins`; keeping it explicit for now avoids build-time magic.
  */
 import type { KhaopadPlugin, PluginInitContext } from "./types";
+// Ensure side-effect module-load registrations (sidebar nav, webhook
+// events) run in the server bundle too. The client bundle triggers the
+// same import chain via sidebar-nav.ts. See registrations.ts for the
+// rationale (single source of truth for the enabled-plugins set).
+import "./registrations";
 
 // ─── Enabled plugins ────────────────────────────────────────
 // Order matters: earlier plugins init first, so their sidebar groups
 // appear above later plugins'. Core groups (registered in
 // $lib/components/admin/sidebar-nav.ts) always come first.
+//
+// Keep this list in sync with the imports in registrations.ts.
 
-import hello from "../../plugins/hello";
+import hello from "$plugins/hello";
 
 const enabledPlugins: KhaopadPlugin[] = [hello];
 
@@ -33,19 +40,24 @@ export async function initPlugins(ctx: PluginInitContext): Promise<void> {
   if (initPromise) return initPromise;
 
   initPromise = (async () => {
-    for (const plugin of enabledPlugins) {
-      try {
-        await plugin.onInit?.(ctx);
-      } catch (err) {
-        // Never crash a request over a plugin init failure. Log and continue.
-        // eslint-disable-next-line no-console
-        console.error(
-          `[plugins] ${plugin.slug} onInit failed:`,
-          err instanceof Error ? err.message : err,
-        );
+    try {
+      for (const plugin of enabledPlugins) {
+        try {
+          await plugin.onInit?.(ctx);
+        } catch (err) {
+          // Never crash a request over a plugin init failure. Log and continue.
+          // eslint-disable-next-line no-console
+          console.error(
+            `[plugins] ${plugin.slug} onInit failed:`,
+            err instanceof Error ? err.message : err,
+          );
+        }
       }
+    } finally {
+      // Mark init done even on unexpected outer throws so we don't
+      // pin every future request to a rejected promise.
+      initialized = true;
     }
-    initialized = true;
   })();
 
   return initPromise;
